Module iterlab::hofs

This module contains some functions that require higher-order functions to implement them.

Implement this module after the iterators module.

We recommend implementing sum_squares before fib_fun, since sum_squares is really just a warmup exercise.

See the function documentation within these modules for more information.

Modules

fib_fun

A lot of documentation, if you choose to read it. Do this second, and only if you have time to spare.

sum_squares

A quick exercise to get familiar with using iterators and higher-order functions. Do this first.