iterlab::hofs

Module fib_fun

A lot of documentation, if you choose to read it. Do this second, and only if you have time to spare.

Functions

read_the_docs

For every ith number from 0..n, find the sum of the first i fibonacci numbers and square that sum.

sum_fib_range

Find the sum of the range of fibonacci numbers defined by start and end.