iterlab::hofs::fib_fun

Function read_the_docs

pub fn read_the_docs(n: usize) -> Vec<usize>

For every ith number from 0..n, find the sum of the first i fibonacci numbers and square that sum.

Then, for each of those squared sums, keep only the ones that are divisible by 2 or 3.

Let’s say that there are now m remaining numbers. For every jth number from 1..m, integer divide the number by j. And since 0 is going to end up being the first of the m numbers anyways, don’t do 0 / 0, and just return 0 instead.

Return these numbers as a vector of usize.

You will probably want to implement sum_fib_range before implementing this.

Example

Let’s walk through a call of read_the_docs(6).

The first 6 fibonacci numbers are:

0, 1, 1, 2, 3, 5

We want 6 summations:

0, 1, 2, 4, 7, 12

Then we square these:

0, 1, 4, 16, 49, 144

We keep only those divisible by 2 or 3 (here only even numbers):

0, 4, 16, 144

Let’s attach the position of each number to the above:

value: 0, 4, 16, 144
index: 0  1  2   3

Dividing each out:

0, 4, 8, 48

Let’s see if this works:

assert_eq!(vec![0, 4, 8, 48], read_the_docs(6));

This time we will not give you any constraints on how to implement this function.

However, you will probably find it a lot easier to use iterators and dot chaining than to use a bunch of for loops. You may even want to reuse some of the code you wrote for previous parts…

Note that our reference solution is 6 short lines (which means we have at least 5 chained methods)!