Function iterlab::hofs::fib_fun::sum_fib_range

pub fn sum_fib_range(start: usize, end: usize) -> usize

Find the sum of the range of fibonacci numbers defined by start and end.

You will probably want to implement this before implementing read_the_docs.

Example:

Recall the fibonacci numbers from this sequence:

0, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144, 233...

If we want to sum the 4th to 8th fibonacci number, we would need to sum 3 + 5 + 8 + 13.

assert_eq!(29, sum_fib_range(4, 8));

Try to not use any for loops, if statements, or mutable variables!

You are also allowed to use code from other parts of this homework. Make sure to use the correct things!

By re-using some code from other modules, our reference solution fits in 1 line!