iterlab::hofs::sum_squares

Function sum_of_squared_odd_numbers

pub fn sum_of_squared_odd_numbers(largest_square: usize) -> usize

Make sure to read the documentation for sum_of_squared_odd_numbers_bad first.

Implement this function to do the same thing as sum_of_squared_odd_numbers_bad, but without any for loops, if statements, or mutable variables!

You could just click the link in the bad version’s docs and copy the answer… but try and do it on your own so you get practice thinking for yourself!

Example:

assert_eq!(35, sum_of_squared_odd_numbers(30));